requirements.txt
# Backend
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.3

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm==0.9.12
numpy==1.26.4
pillow==10.2.0
opencv-python==4.9.0.80

# Utilities
pyyaml==6.0.1
tqdm==4.66.2
scikit-learn==1.4.1.post1
matplotlib==3.8.3
